[7.1.2] Update `Lockfile` to implement `Read`/`Write` and use `std::io::Error`
This patch updates `Lockfile` to implement the `Read`/`Write` traits so
that it can be used by the newly generic `Checksum<T>` that now works
with `File` and `Lockfile`. This also refactors `Lockfile` away from
`anyhow::Error` to `std::io::Error` because `std::io::Result` is
required by those traits and downcasting things is a mess.