No dependency injection for `stdout` / `stderr`
This patch removes the dependency injection for `stdout` / `stderr`
because it won't be needed by tests, since the tests have to be run in
subprocesses using `assert_cmd`, which provides its own helpers for
testing stdout/stderr. I left `stdin` injected for now because I'm not
sure how or if that will be used in tests.