[8.4.1] Extract common command setup into `CommandContext`
This patch deviates from the book slightly and introduces
`CommandContext` to extract the common repository setup. I tried
following the structure in the book and adding `repo` as an attribute on
a new `Command` trait, but default implementations on traits can only
access other methods, not attributes. Instead, `CommandContext` contains
all of the shared attributes.