Move `jit add` tests to integration tests
It turns out Cargo only builds the `jit` binary for integration tests,
not unit tests. This means that the unit tests were running on the last
result of `cargo build` and not necessarily the current code. This patch
restructures `jit` to be a library and binary crate and moves the `jit
add` tests to the new integration tests.