Refactor `Status` and `Init` commands to use `self`
This patch refactors the `Status` and `Init` commands to use `self` to
store attributes of the `CommandContext`. This is because 9.2.1, Changed
contents, adds a bunch of attributes to `Status` that will be easier to
work with on self rather than passing them around to various functions.

I started refactoring `Add` and `Commit` as well but was thwarted by the
compiler and lost interest.