[10.2.1] Detect files added to the index
With our newfound ability to read objects from the database, we can
detect and report differences between the HEAD tree and the index. Since
we'll need to check on every file in the tree, we include a recursive
procedure to load the entire tree and store a mapping from pathnames to
tree items that includes the file's mode and blob ID.

The first change we detect is that a file has been added, that is, it
exists in the index but not in HEAD. This is simple enough to detect, we
just need to adjust our reporting code to recognise the new
`index_added` status and report index changes with an `A` in the first
column.