[10.2.2] Detect files modified in the index
Following from the previous commit, this is a very minor change. If we
find a corresponding tree item for an index entry, and its mode or blob
ID differs from the values in the index, then that file has been
modified in the index. We record this with the `IndexModified` state
and list it with status `M`.