[10.2.3] Detect files deleted from the index
The final index-related change we must detect is that a file exists in
the HEAD tree but not in the index, indicating that it's been deleted.
By deleting files from our copy of the tree as we scan through the
index, we know that any tree items left at the end must not exist in the
index, and we mark these with the `IndexDeleted` state and print them
with a `D`.