[12.1] Extract status logic into `Repository`
This sets us up to work on the `diff` command. In order to know which
files we should calculate and display the diffs of, without reading
every file in the tree, we need to know which files differ between the
HEAD commit, the index, and the workspace. That is, we need exactly the
information that the `status` command displays.

Here we extract the business logic from `Status` into `Repository` so
that it can be reused by other commands. The moved code is kept pretty
much verbatim, and the only code left in `Status` is that related to
printing the output, and opening and writing the index. In the book,
they create a separate `Repository::Status` class that has a reference
to its parent `Repository` and this quickly became a mess in Rust, so I
put all of its methods on `Repository` instead.

The index logic should remain in the top-level command code rather than
moving into reused business logic, because it's a piece of global state.
If the top-level command has opened the index for updates and a piece of
business logic then attempts the same thing, the second call will fail.
Therefore we leave top-level commands in charge of how the index is
opened.