[12.3] Display line-by-line diffs for changed files
Having completed the functionality to show the headers for all
combinations of HEAD/index/workspace changes, we can now augment the
`diff` command with the ability to display a line-by-line patch for each
file.

We can use the diff module we introduced earlier for this. It just
requires that each `Target` object exports a `data` field for the
contents of the blob it represents. HEAD and index targets read a blob
from the database, while workspace targets return the file contents that
was used to compute the new uncommitted object ID.