Inject `stdout` and `stderr` into `CommandContext`
This patch reverses 0c0460b254895901e38d85d6eb1f0731aabd9d4a and injects
stdout and stderr into `CommandContext`. This is a prerequisite for
adding pager support to `jit` because Rust doesn't provide a way to
redirect stdout/stderr as used by `println!()` and `eprintln!()`, so we
have to use `writeln!(stdout, ...)` and `writeln!(stderr, ...)` instead.

`writeln!()` requires its buffer to be mutable which means that if we
use `writeln!(self.ctx.stdout, ...)` or
`writeln!(self.ctx.stderr, ...)`, every method that writes to
stdout/stderr has to borrow `&mut self`, which means that we can't
iterate over e.g. `self.ctx.repo.workspace_changes` and call a method
that prints output in the body of the loop. This is incredibly annoying,
so we use `RefCell<T>` and interior mutability to enable those methods
to only borrow `&self`.