[12.3.3] Run a pager for displaying the `diff` output
To let the user navigate the potentially long output of `diff`, we
display it by running a pager like less(1) as a child process, and
piping the command's output into the pager.

We maintain colour formatting output even though we're no longer writing
to a TTY; less(1) is run with the `R` flag and lv(1) is run with `-c` so
that they print colour-formatted content without escaping it.