[13.2] Create branches from the current HEAD
This commit introduces the ability to create branches, which are
represented as pointers to commits, just like `.git/HEAD`. They're
stored by name in the file `.git/refs/heads/<name>`.

According to comments in the Git source [1], a branch is invalid if:

    - any path component of it begins with ".", or
    - it has double dots "..", or
    - it has ASCII control characters, or
    - it has ":", "?", "[", "\", "^", "~", SP, or TAB anywhere, or
    - it has "*" anywhere unless REFNAME_REFSPEC_PATTERN is set, or
    - it ends with a "/", or
    - it ends with ".lock", or
    - it contains a "@{" portion

We represent these rules in the `INVALID_NAME` regex.

The implementation in the `Refs` struct generalises the logic for
reading and writing `.git/HEAD` to arbitrary references, searching the
`.git`, `.git/refs`, and `.git/refs/heads` directories for the given
name. This means the same file-locking technique that's been used for
writing HEAD is also used for writing branch pointers.

[1]: https://github.com/git/git/blob/v2.15.0/refs.c#L58-L70