[13.3.1] Parse revision notation
Often we'd like to create a new branch pointing at something other than
the current HEAD. While strictly speaking it is possible to do this
using commit IDs, it is far from convenient, particularly as we don't
yet have a `log` command or any other tooling for viewing the commit
history, other than by manually inspecting the object database.

Git includes a small query language for selecting commits, which it
calls revisions [1]. Here we implement a parser for a small subset of
this notation:

- The `^` operator selects the parent of a commit
- The `~<n>` operator selects the nth ancestor, e.g. `~3` == `^^^`
- `@` is an alias for HEAD
- All other strings not matching `INVALID_NAME` are references

We don't yet interpret these strings, we just parse them into a
structure for recognition, which can then be further processed.

[1]: https://git-scm.com/docs/gitrevisions