[13.3.2] Set the branch start point using a revision
Given the revision parser in the previous commit, we now allow the
`branch` command to set a starting value for the branch other than the
current HEAD. The `Revision` struct is expanded to include an
interpreter for revision strings, which walks the parse tree and applies
the necessary operations.

If no argument is provided, the current value of HEAD is used as the
start point. If the revision does not resolve to an object, an error is
raised.