[13.3.3] Handle short object IDs in the revision resolver
Git's revision notation allows an object to be identified using a prefix
of its full ID, as long as that prefix uniquely identifies it. As long
as the prefix is at least four characters, Git will attempt to find a
unique object matching it.

To support this, we add the method `Database.prefix_match()`, which
takes a partial ID and returns a list of full IDs matching it, by
inspecting the files in `.git/objects` and selecting those filenames
that match.

The `Revision` struct uses this as follows. To interpret a reference, it
first tries to find it using `Refs`, as before. If that yields nothing,
we attempt to interpret the string as a partial ID. If this yields
exactly one result, we return it, otherwise the search fails.