[14.2] Plan workspace changes on checkout
Here we begin an implementation of the `checkout` command by determining
the changes needed to make the workspace reflect the contents of the
checked-out commit.

We need to plan how to modify the workspace to migrate its current state
to that of the target commit. The order in which these changes are
applied is important; a file that exists in the current commit may be
blocking the creation of a file from the target commit, and so deletions
must be processed before other types of changes.

The `Migration` struct collates the tree diff entries into three
categories: creations, updates and deletions, where each category is a
list of pairs of a `PathBuf` and a `Entry` for the new content that
should be written to that path. For deletions, the entry is always
`None`.

    {
      Action::Create: [(PathBuf, Entry)],
      Action::Update: [(PathBuf, Entry)],
      Action::Delete: [(PathBuf, Entry)]
    }

We also build two sets of directories: `rmdirs`, which records those
directories which it may be possible to delete after the deletions are
effected; and `mkdirs`, those directories that must exist before newly
created files can be written.