[14.5.1] Refactor `Repository` to support single-file status checks
The `checkout` command attempts to preserve uncommitted changes in the
index and workspace when switching between commits. It does this by
working out which files differ between the current HEAD and the target,
and only updating those files, rather than replacing the entire tree.

However, sometimes the index and workspace may contain changes to files
that would be affected by the checkout. In such cases, we should detect
that the checkout would overwrite uncommitted data and refuse to go
ahead.

To do this, we need to reuse some of the logic from the `status` command
that compares the HEAD tree, index and workspace and works out what's
changed. Rather than performing a full scan of the repository, we just
want to check those files the checkout would modify. To support this,
we're breaking up `Repository.check_index_against_workspace()` and
`Repository.check_index_against_head_tree()` into:

- `compare_index_to_workspace` compares an index entry to a filestat
  from the workspace. It first compares them by size, mode, and
  timestamps, and finally by hashing the file. It does not update the
  timestamps in the index if the file is unchanged; that logic remains
  in `Repository::Status`.

- `compare_tree_to_index` compares an entry from a tree to one from the
  index, comparing their modes and object IDs.

In the book, this is accomplished by creating a `Repository::Inspector`
class and extracting its functionality from `Repository::Status`.
Unfortunately I haven't figured out how to create the circular
parent/child struct relationship in Rust used by these classes, so I
just added the new methods to `Repository` like in
47066350ce04789c9a2e3344b606a571c07ee4b5.