[15.2, 15.2.1] troduce symbolic references for checking out branches
Say we have the following history, where HEAD (`.git/HEAD`) and `foo`
(`.git/refs/heads/foo`) point at the latest commit, and `bar`
(`.git/refs/heads/bar`) points at HEAD's grandparent.

    A       B       C       D
    o <---- o <---- o <---- o
	    |               |
	  [bar]           [foo]
			  [HEAD]

`foo` and `bar` were created using the `branch` command. Currently, if
we run `checkout bar`, the index and workspace are migrated and HEAD is
updated so that it points at the same commit as `bar`.

    A       B       C       D
    o <---- o <---- o <---- o
	    |               |
	  [bar]           [foo]
	  [HEAD]

However, this doesn't result in the `bar` branch growing when we make
new commits. The `commit` command updates HEAD to point to the new
commit, but `bar` remains unchanged. HEAD and `bar` are not connected.

    A       B       C       D
    o <---- o <---- o <---- o
	   / \              |
	[bar] \           [foo]
	       \
		\ X
		 o
		 |
	       [HEAD]

As we add more commits, HEAD drifts away from `bar`, whereas we'd rather
that the `bar` branch pointer followed along so that we retain a
reference to this line of history after we check out some other commit
and HEAD no longer points at it. That is, instead of this:

    A       B       C       D
    o <---- o <---- o <---- o
	   / \              |
	[bar] \           [foo]
	       \
		\ X      Y       Z
		 o <---- o <---- o
				 |
			       [HEAD]

We want this:

    A       B       C       D
    o <---- o <---- o <---- o
	     \              |
	      \           [foo]
	       \
		\ X      Y       Z
		 o <---- o <---- o
				 |
			       [bar] <== [HEAD]

That `<==` arrow represents a symbolic reference or symref. When we run
`checkout bar`, instead of writing the new commit ID directly to
`.git/HEAD`, we can instead write this:

    ref: refs/heads/bar

HEAD is now a symref to `bar`, and any operation that wants to read or
update HEAD should instead update whatever branch reference HEAD is
pointing at.

In this commit we introduce the `Refs.set_head()` method, which takes an
object ID and the original unresolved revision name that `checkout` was
run with. If that name is the name of a branch, then `Refs` will write a
symref to `.git/HEAD`, otherwise it writes the object ID as usual. We
also update the `read_head` and `read_ref` methods so that they traverse
symrefs, and introduce the `current_ref` method which returns the
identify of the reference that HEAD currently points at.