[15.5.1] Introduce StructOpt to parse command-line options
So far, Jit makes very limited use of command-line option switches.
There's the `status --porcelain` option for generating a
machine-readable status output, and `diff --cached` option for showing
differences between HEAD and the index. For the `branch` command, we're
about to introduce the `--verbose`, `--delete` and `--force` options, as
well as the shortands `-v`, `-d`, `-f` and `-D`, and some of these
options will be intermingled with normal positional arguments.

To handle this, we're introducing
[StructOpt](https://github.com/TeXitoi/structopt). I don't love it. I
might switch to [`clap`](https://github.com/clap-rs/clap), but that is
stagnating on v2 while waiting years for a v3 release. Ugh.