Read `Commit` `oid`s from disk
Previously we would always recalculate the `oid` in `Commit.oid()` which
would be wrong if the commit had been amended, because we don't use the
`Committer` line in calculating the `oid`. I noticed this after amending
the previous commit ([16.1.4] Decorate log entries with the refs that
point to them) when `jit log` wouldn't decorate that commit with
`HEAD -> main`. This is because it was using the calculated, non-amended
`oid` and so not matching the `oid` of the ref.

In this patch we update `Commit::parse()` to use the `oid` we read from
disk as the `Commit`'s `oid` and fix this issue.

It might be worth additional refactoring to match the structure in the
book, where `Object.oid` is calculated once and set during
`Database.store()` or read from disk in `Database.read_object()`.