[16.1.5] Extract printing logic from `commands::diff::Diff`
We'd like to add support for the `--patch` option to the `log` command,
which prints diffs for all the files changed in a commit. This logic
already exists in the `commands::diff::Diff` class, we just need to
extract it.

This commit extracts all the logic related to printing diffs into a
shared module called `commands::shared::print_diff::PrintDiff`. It
includes the `Target` class that carries the path, object ID, mode and
data for each diff input, but not the methods for constructing them
since that's specific to each command.