Fix sorting of sub-trees in `Tree` serialization
This patch fixes an edge case in `Tree.bytes()` to sort entries in the
order expected by Git. Given:

    foo.txt  # a file
    foo:txt  # a file
    foo      # a directory

Git expects the above to be sorted like:

    foo.txt  # a file
    foo      # a directory
    foo:txt  # a file

Basically, it sorts directories in the tree as if they have a trailing
slash in their name, e.g. `foo/`.

I discovered this when I tried to take advantage of the Rust 2018 module
layout, e.g.:

    src/commands/shared.rs
    src/commands/shared/print_diff.rs

and was unable to pull the repo after committing with `jit`:

    error: object 9c62b839a344594563221bb271be3ee5ce30f426: treeNotSorted: not properly sorted
    fatal: fsck error in packed object
    fatal: index-pack failed