[16.2] Start logs at an arbitrary revision
The last few commits focus on how the `log` command displays information
about each commit. We now shift focus to look at how this command
selects which commits it will display.

The first adjustment we can make here is to let `log` take an argument
that, if present, causes logging to commence at the named commit rather
than at HEAD. This commit name is resolved using the `Revision` class
that we originally developed as a way to identify the starting position
of a branch pointer.