[16.2.2] Logs with multiple start points
Until now the `log` command supported following a chain of commits from
a single starting point. But now suppose we have the following history;
`parent` links point from right to left, the `main` branch points at
commit D and `topic` at commit G.

    A   B   C   D
    o---o---o---o [main]
	 \
	  o---o---o [topic]
	  E   F   G

If we run `log main topic`, we should see a list of every commit in
the project where each commit appears exactly once. In particular, we
should not see a list of all the commits reachable from `main`, and
then all the commits reachable from `topic` -- this would present a
confusing chronology and would print commits A and B twice.

Instead, we'd like to display the histories of `main` and `topic` back
to the point they diverged, followed by their common history. So, an
order like this would be acceptable:

    D, C, G, F, E, B, A

The commits {D, C} must appear in that order because they are directly
linked by `parent` pointers. Similarly {G, F, E} should be ordered as
such. However, there is no fixed order between commits on either branch,
e.g. C is not linked to E via `parent` pointers and so C and E are not
mutually ordered. All these commits should however appear before B,
since B is an ancestor of all of them.

`RevList` achieves this using a graph search algorithm. The two starting
commits D and G are placed in a date-ordered queue. We consume commits
from this queue, yielding them to the caller if they've not yet been
visited. We then examine their parents and add them to the queue. A set
of flags is stored against each commit to mark whether it's been
processed and its parents followed.

Commits are yielded to the caller as soon as they are found, so we do
not need to load the entire history into memory before printing any
output. Instead, output begins as soon as the first commit is loaded.

Assuming users' clocks are roughly accurate, each commit will usually
have a date that's later than its parent, so prioritising the commit
queue by date will usually make sure that all descendants of B are
processed before B itself. So, regardless of the length of the two
branches, B should be correctly ordered relative to any later commit.