Fix flaky tests
This patch attempts to fix the flaky `log_test`s by calling a function
to get the commits instead of using a fixture. I suspect there's some
weird interaction going on when both the `commits` fixture and the test
requested the `helper` fixture. Maybe it was being run twice? I don't
particularly feel like debugging that right now and this change seems to
fix the flakiness.