[16.2.3] Exclude branches from logs
Given the following history, where `parent` links point from right to
left, the `main` branch points at commit D and `topic` at G:

    A   B   C   D
    o---o---o---o [main]
	 \
	  o---o---o [topic]
	  E   F   G

It would be useful to be able to see just those commits that are on the
`topic` branch, before merging said branch into `main`. By this we mean
those commits that are reachable from `topic` but not from `main` --
commits that make changes that are not included in the current state of
`main`.

Git supports two notations for this set of commits: `main..topic` and
`^main topic`. Both of these notations indicate a commit list that
includes commits reachable from `topic` but not from `main`, that is
this set:

	  o---o---o [topic]
	  E   F   G

To generate this set, `RevList` proceeds much as before, using a
date-ordered queue to process commits from the given starting points.
The twist is that any excluded start point -- one prefixed with `^` or
on the left of a `..` operator -- gains the flag `uninteresting`, and
that flag is propagated to its parents. Commits marked with this flag
are excluded from the results.

The one problem is that when we have excluded start points we cannot
yield commits as soon as they are found. Until we have found the commit
where `main` and `topic` intersect, we don't know where that commit will
be -- it could be any commit on `main` we've already seen.  Therefore in
this mode, we need to find all the included commits first, and then
display them. The cost of this is alleviated by trying to stop searching
as soon as possible; rather than scanning back to the root commit, we
stop as soon as all the commits in the queue are uninteresting.