[17.3] Allow commits to have multiple parents
In order to write merge commits, we need to be able to say that a commit
has more than one parent. To support this, a number of changes are
necessary, plus a little facade work to keep existing code working.

First, the `database::commit::Commit` constructor now takes a vector of
parents, rather than a single object ID. It writes one `parent` header
to disk for every parent it has.

Second, `Commit.parse()` must be modified so that it can handle multiple
headers with the same name. For most headers it just uses the first
value it saw, but for `parent` it uses all the values to construct the
resulting `Commit` object.

To keep existing code working, notably `CommonAncestors`,
`RevList` and the `<rev>^` and `<rev>~<n>` notations, the method
`Commit.parent()` must continue to work. In a multi-parent history, the
revision notations specifically refer to the first parent, so we can
make `Commit.parent()` return the first parent to keep those working.
This change will also keep `CommonAncestors` and `RevList` working for
existing histories, but they'll be incomplete when we need to begin
processing histories containing merges.