[17.3] Extract `WriteCommit` from `Command::Commit`
In our `merge` command, we're going to want to write a commit. We're
therefore extracting the logic for doing this from
`commands::commit::Commit` into a new struct, `WriteCommit`.

This struct takes care of the following concerns:

- Reading the index and saving its tree to the database
- Reading the author information from the environment
- Writing a commit object from this tree, author, and the given args
- Updating HEAD with the new commit's ID

This method does not know how to set the parent(s) for the commit -- it
takes the parents as an argument so each command can decide them for
itself.