[17.4] Basic `merge` command: no conflicts, no unstaged changes
Having added the infrastructure for finding the common ancestor and
writing a merge commit over the last few commits, we can now add a basic
version of the `merge` command. Its working is fairly straightforward
piecing-together of existing functions.

First we read HEAD and resolve the input argument to a commit ID --
these are the tips of the merge. Then we find the common ancestor of
these tips. From this, we work out the tree diff of the merge target
against the common ancestor, and use a migration to apply that
difference to the index and workspace. Finally we dump the index to disk
and write a commit from it, using the two tips as the new commit's
parents.