[17.5] Find the best common ancestor with merges
Now that we can create merge commits, we need to make our
`CommonAncestors` struct a bit more sophisticated. Say we have the
following history, in which `parent` links point from right to left.

	D               A
    o---o---o---o---o---o
	 \     /
	  o---o---o
	      C   B

A "common ancestor" of two commits is any commit that's reachable from
both of them. Commits C and D are both common ancestors of A and B.  A
"best common ancestor" is one that is not an ancestor of any other; D is
an ancestor of C, and so C is the best common ancestor of A and B.

Our current `CommonAncestors` implementation only follows the first
`parent` link from each commit, so it will never discover that C is an
ancestor of A. This is the first thing we must fix.

That's not all. We currently scan backward from A and B and return the
first common ancestor we find, but this will no longer suffice. Consider
this history:

	D               A
    o---o---o-------o---o
	 \         /
	C o---o---o
	   \
	    o---o
		B

Commit D is 3 steps away from A and B, whereas C is four steps from A.
So, just depending on the order commits are visited in, we may decide D
is the best common ancestor here, when in fact it is C.

This commit implements a partial solution to this problem. Rather than
returning the first common ancestor we find, we keep a list. When a
commit is found with both `parent1` and `parent2` flags, it's added to
the results, and its parents are marked `stale`. We search until all
commits in the queue are marked `stale`, and then return any results
that are not marked `stale`.