[17.5] Find the base commits for a merge
We currently have the `CommonAncestors` struct for finding the best
common ancestors of two commits. In some cases, it can stop searching
before it identifies that one of the results it's found is an ancestor
of another, and is therefore not a best common ancestor. To fix this, we
introduce another struct that runs `CommonAncestors` for every pair of
results returned by the first search, to filter out bad candidates.

	D               A
    o---o---o-------o---o
	 \         /
	C o---o---o
	   \
	    o---o
		B

For example, imagine that when we run `CommonAncestors` on the above
history for commits A and B, it returns commits C and D. To identify the
best common ancestors, we run `CommonAncestors` again with C and D as
inputs, and when it's done searching, we check the flags that have been
added. We should see that D is flagged as `parent1`, and is thus an
ancestor of C, and should be removed from the results.

The `Bases` struct performs this filtering, by running `CommonAncestors`
with the original inputs, and then again with every result commit
against all the other commits. Those found to be mutual ancestors are
filtered out.