[17.6.1] Log all commits reachable from merge commits
In the `log` command, based on the `RevList` struct, we explore the
commit graph back from the given starting point and print every commit
we find. However, the implementation only follows the first `parent`
link from each commit. Consider this history:

    A   B   C   G   H
    o---o---o---o---o
	 \     /
	  o---o---o
	  D   E   F

`RevList` should discover the following commits if it's started at
commit H:

    H, G, C, E, D, B, A

However, because we only follow the first parent, commits D and E will
be excluded from the results. This can be fixed by iterating over all
the parents of a commit, rather than only looking at the first one, when
adding a commit's parents to the input queue.