[17.6.1] Exclude all commit parents from log results
When a commit is given as an excluded input to `RevList`, using the
`A..B` or `^A` notations, then everything reachable from that commit
should be excluded. Consider this history:

    A   B   C   G   H
    o---o---o---o---o
	 \     /
	  o---o---o
	  D   E   F

If we run `log H..F`, we should see only commit F in the output.
However, because we only follow the first parent when propagating the
`uninteresting` flag, commits D and E will also appear.

We resolve this by following all parents of commits when marking commits
as uninteresting. This is done using a queue and a loop rather than
recursion, since the history can become very large and cause a stack
overflow if using a recursive implementation.