[18.2.2] Detect fast-forward merges and update HEAD
If the history looks as below, and we have `main` checked out, then
running `merge topic` does not require us to create a merge commit.

A   B
o---o [main]
     \
      o---o [topic]
      C   D

That's because this history is not divergent, it's equivalent to

A   B   C   D
o---o---o---o

Because `main` is the best common ancestor for this merge, nothing has
changed on the `main` branch that could need to be reconciled with the
other branch. So, we can get the correct final state by just
"fast-forwarding" the `main` branch pointer to point at the target
commit.

This edge case can be detected by checking whether the best common
ancestor is the same as the current HEAD.