[18.3.3] Key index entries by stage
To represent that a file is conflicted following a merge, Git uses two
bits in the `flags` field of each index entry [1]:

    A 16-bit 'flags' field split into (high to low bits)

      1-bit assume-valid flag
      1-bit extended flag (must be zero in version 2)
      2-bit stage (during merge)
      12-bit name length if the length is less than 0xFFF; otherwise 0xFFF
      is stored in this field.

The `stage` is a number from 0 to 3 that models conflict status. In
normal operation, each pathname appears only once in the index and has
stage 0. During a merge conflict, a pathname may be indexed multiple
times, each entry having a different non-zero stage. The object ID of
each entry by stage is:

1: the version of the file from the merge base commit
2: "our" version, that is the version from HEAD
3: "their" version, that is the version from the merge target commit

Before we start setting stages on entries, we need the index to be able
to hold multiple entries for the same path, and we do this by switching
`Index::Entry#key` from the entry's path, to the entry's path and stage.
Ruby's sorting methods will still keep these array-valued keys in order.

The stage is derived from `flags` using some bitwise arithmetic to
select the 3rd and 4th most-significant bits.

A few changes are necessary to the `Index` methods for locating and
removing entries -- either defaulting the stage to 0, or performing an
operation for all possible stage values.

[1]: https://git-scm.com/docs/index-format