[18.4.3] Detect file-directory conflicts during merge
The previous commit handles cases where both sides of a merge modify the
same path. It's also possible for both sides to modify different paths
in a way that makes it impossible to combine them into a single tree.

For example, if one side adds the file `f.txt`, while the other adds
`f.txt/g.txt`, it is impossible to merge both these since `f.txt` must
be either a file or a directory, not both. In this case, we keep the
innermost path `f.txt/g.txt` in the index and workspace, and mark the
outer path `f.txt` as conflicted. We write the copy of `f.txt` with a
modified path, e.g. `f.txt~HEAD`, resulting in an untracked file.