[19.3.1] Show unmerged paths in `diff` output
If `Repository` reports that a certain file is conflicted, we can't take
a normal diff for that file because there may be more than one entry for
it in the index.

For now, we'll solve this by simply printing that the path is unmerged,
rather than trying to do anything more complex.