[19.4.2] Persist pending merge commit state to disk
If a merge is conflicted, we need to exit and let the user fix the
conflicts before making a commit. But, writing the merge commit requires
the target object ID and message the user originally passed to `merge`.

To retain this information, we introduce a new class called
`PendingCommit` which saves this data in `.git/MERGE_HEAD` and
`.git/MERGE_MSG`. The presence of these files indicates a merge is in
progress.