[20.1.2] The diff3 merge algorithm
Here we implement an algorithm for merging a file when both sides of a
merge have modified it. This can still result in conflicts, but if each
side's changes do not overlap, they can be merged cleanly. When
conflicts do occur, they are isolated to just the conflicting region
rather than the entire file.

The algorithm is that of diff3 [1], a command-line tool whose algorithm
is explored in a paper by Khanna, Kunal and Pierce [2].

For the diff function, we use our existing `Diff` module, so
`Merge::Diff3` works on similar data representations: strings that are
split by line, or lists of arbitrary values.

[1]: https://manpages.ubuntu.com/manpages/bionic/en/man1/diff3.1.html
[2]: https://www.cis.upenn.edu/~bcpierce/papers/diff3-short.pdf