Update `database::tree::Tree` and `database::entry::Entry` to better match the book
In the book, a `database::entry::Entry` only has an `oid` and a `mode`,
no `name` and `database::tree::Tree::build()` takes an array of
`index::Entry`, not `database::entry::Entry`. This patch refactors
`database::tree::Tree` and `database::entry::Entry` to match that
structure, because in 21.1.1 the book starts creating
`database::entry::Entry`s without a `name`.