[21.1.1] Report an error when removing an untracked file
If an argument passed to `rm` does not match anything stored in the
index, this is most likely a typo on the part of the user, or a logical
error in a script they're running that has wrongly assumed the file
exists.

Therefore if any argument to `rm` is not the name of a tracked file, we
exit with an error.