[21.1.2] Remove files from index, not from the workspace
The `rm` command takes an option called `--cached`, which causes it to
remove a file only from the index, not from the workspace. It still
tries to avoid losing data, and this command will fail if the index
entry differs from *both* the HEAD item and the file on disk. If the
entry is equal to either of them then we're not losing anything by
removing it.