[21.1.2] Recursively remove files from the index
As well as names of single files, `rm` accepts directory names. When
given an argument that names a directory (as stored in the index, not
the workspace), we can find all the child paths under that directory in
the index and remove all of them.

So, when the `-r` option is given, each argument is expanded so that
directory names are replaced with all their descendant paths, and then
this list of paths is checked for changes as usual. If `-r` is not
given, it is an error to ask to remove a directory.

To support this behaviour, `Index` has gained a couple of methods for
extracting the child paths of a directory, and finding out if something
is a tracked directory, as opposed to a file. The `Workspace.remove()`
method will now ascend up the parent directories of the removed file and
attempt to delete them, in case they have become empty as a result of
that file being deleted.