[21.2.1] Reset index entries to a particular commit
Rather than just resetting index entries to the version from HEAD,
`reset` allows the first argument to identify a commit from which to
take the file. In Git, the arguments are ambiguous -- the first argument
could be a pathname or a revision, and it guesses which it is by trying
to interpret it as one and then the other. If arguments are placed after
`--` in the command list, Git knows they're definitely pathnames.

However, clap/structopt doesn't seem to differentiate between arguments
appearing before and after `--`; it just uses this word to stop
interpreting arguments as option names. This could have been solved much
more cleanly in Git by requiring an option name to identify the revision
argument if present and assuming all positional arguments are filenames,
but alas. We'll have to do our best to guess instead.