[21.3] Add the `--soft` and `--mixed` modes to `reset`
When run with a list of pathnames, `reset` resets those paths in the
index to their version in the given commit or HEAD, and does nothing
else -- it does not touch HEAD or the workspace.

The other main behaviour of `reset` is to take a mode rather than a list
of files, where the modes work as follows, `mixed` being the default.

- `soft`: move HEAD so that it points at the given commit

- `mixed`: as `soft`, plus restore the index to contain the contents of
  the given commit

- `hard`: as `mixed`, plus restore the workspace to match the index and
  HEAD as of the given commit

If no commit is given, the current value of HEAD is used. This means
`reset [--mixed]` can be used to effectively unstage all changes by
restoring the index to match HEAD, and `reset --hard` can be used to
throw away all unstaged/uncommitted changes.

In contrast with `checkout`, when `reset` updates HEAD it moves the
branch that HEAD is pointing at. For example, given this starting state,
where HEAD is a symref to the `master` branch:

    o---o---o---o---o---o---o
			    |
			[master] <== [HEAD]

Running `checkout @~3` writes to `.git/HEAD` itself so that it points at
the given commit, leaving `master` alone.

    $ jit checkout @~3

    o---o---o---o---o---o---o
		|           |
	     [HEAD]     [master]

However, `reset` moves the branch that HEAD is pointing at, just like
`commit` does. So `reset @~3` moves the `master` branch to the given
commit and leaves HEAD pointing at the branch pointer.

    $ jit reset @~3

    o---o---o---o---o---o---o
		|
	    [master] <== [HEAD]

This means `reset` can be used to "remove" commits from the tips of
branches. The commits still exist, but they're no longer reachable via
the moved branch head.

In this commit we just implement the `--soft` and `--mixed` modes as
they require fairly minor changes. We update HEAD as long as no
filenames were given, and we clear out and replace the index contents
from the given commit's tree if we're not in `soft` mode.