Extract status logic into `repository::status::Status`
In this patch we extract logic from `repository::Repository` into
`repository::status::Status`. In the book, this was done when working on
the `diff` command in 4536d956a32f9b39aa6e1cf490df65d032b16df8 but I
just added those methods on `repository::Repository` because the
relationship between `commands::status::Status`,
`commands::CommandContext`, `repository::Repository`, and
`repository::status::Status` was difficult to model in Rust.
Unfortunately, the next section 20.3.1 modifies the `Status` class to
load the tree from arbitrary commits which doesn't make sense in our
version where the methods are on `Repository`.

This patch uses a `Box` and raw pointer to the `Repository` struct to
make the `Status` struct possible. This also adds a few `unsafe` blocks
to the codebase.