[21.4] Abort partially completed merges
If a merge gets into a bad state we'd like to be able to escape to the
state before the merge began. This is now possible as we have
`Repository.hard_reset()`.

Running `merge` to begin with writes the current HEAD ID to ORIG_HEAD.
During a conflict, running `merge --abort` will recall the commit ID
from ORIG_HEAD, remove the `.git/MERGE_{HEAD,MSG}` files created by
`PendingCommit`, perform a hard reset on the index and workspace, and
finally update HEAD to point at its pre-merge position.