[22.1] Set commit/merge messages using `--message` or `--file`
Setting the commit message is currently done by reading from stdin.
While it's a workable way of composing multi-line messages, to write
them in a file and then run `cat <file> | commit`, it's hardly
convenient. As a stepping-stone on the way to built-in editor support,
here we add two new options to `commit` and `merge` for passing in the
message:

- `-m`, `--message` allows the message to be written directly on the
  command-line, which is handy for single-line messages

- `-F`, `--file` lets the message by read from a file, which is more
  suitable for composing multi-line messages or taking input from other
  sources