[22.2, 22.2.1] Launch an editor for preparing commit messages
The workflow for writing commits and merges can be improved by opening a
text editor to let the user input the message. This is particularly
useful for resuming merges, when the user might want to amend the
message after having fixed up the conflicts. This will also let us add
the `--amend` option for `commit`.

The main addition here is the `Editor` struct, which takes a pathname,
an environment hash, and a function:

    Editor.edit(path, env, |&editor| {
	# ...
    })

This opens the given file for writing, and then yields to the block,
which can write content to the file. `puts` writes content verbatim, and
`note` writes content with a `#` before each line. Once the block exits,
`Editor` launches a text editor using `GIT_EDITOR`, `VISUAL` or `EDITOR`
from the environment. When the text editor program exits, `Editor.edit`
returns the final content of the file, with comments stripped out.

This `Editor` struct is used to implement the following use cases. Even
in cases where the user's text editor is not opened, the relevant
content is still written to the edited file.

On `commit`, unless `--message` or `--no-edit` is used, the editor is
opened with some instructions for adding a message to the commit.