[22.2.2] Launch an editor for perpraing merge commit messages
This continues the previous patch by implementing additional use cases
with the `Editor` struct:

On successful `merge`, unless `--message` or `--no-edit` is used, the
editor is opened with instructions for writing a merge message. The
message defaults to `Merge commit '<rev>'`.

On a conflicted `merge`, the given message (`--message` or `--file`) or
the default merge message is written to the `.git/MERGE_MSG` file, with
a list of the conflicted files in the comments. The new method
`Index.conflict_paths()` provides this list of files.

When resuming from a merge conflict to write the merge commit, the
message is retrieved from `.git/MERGE_MSG` and some comments reminding
the user they're writing a merge commit are added. The text editor is
then launched and the result becomes the merge commit message.