[22.3.1] Amend the existing HEAD commit
The `commit --amend` operation provides a shortcut for the common
workflow for replacing the current HEAD commit with a modified version:

    $ jit reset --soft @^
    # make changes, amend the index
    $ jit commit -c ORIG_HEAD

When `commit` is run with the `--amend` option, it writes a commit from
the index as usual. But, instead of using the current HEAD as the
parent, it uses HEAD's *parent*. It also pulls the message from the
current HEAD commit and lets you edit it before committing.

So, if we begin in the following state:

	  A   B   C   D
    . . . o---o---o---o
		      |
		   [HEAD]

Running `commit` as normal just pushes a new commit onto this chain:

	  A   B   C   D   E
    . . . o---o---o---o---o
			  |
		       [HEAD]

However, running `commit --amend` results in this situation:

	  A   B   C   D
    . . . o---o---o---o
		   \
		    o E'
		    |
		 [HEAD]

Given the same index state, both commits E and E' contain the same
content. The only difference is the way the message is prepared, and the
fact that the `parent` pointer "skips" the current HEAD. This makes it
appear as though commit D has been removed. Commit D still exists in the
database, it's just no longer reachable from HEAD. You're not genuinely
amending a commit; changing a commit means it gets a new ID and becomes
a new object.

Despite the fact that `--amend` potentially makes commit D unreachable,
it does not set ORIG_HEAD.