[23.1] Cherry-pick a single non-conflicting commit
Cherry-picking involves transplanting the change introduced by one
commit onto a different parent. For example, given this history:

    A   B   C
    o---o---o [main]
	 \
	  o---o
	  D   E

If we cherry-pick E onto `main`, that means generating a commit that
applies the difference between D and E, on top of the commit C.

    A   B   C   E'
    o---o---o---o [main]
	 \
	  o---o
	  D   E

This is equivalent to performing a three-way merge between C and E, with
D as a base. Treating it this way means the difference from D to E will
be applied to the index state of C to generate the new commit, but if
the changes between D and C overlap with this diff, we'll get a
conflict.

We can reuse the `merge::resolve::Resolve` struct to perform this merge,
but we don't want to use `merge::inputs::Inputs` to select the base
commit, as that would select commit B. Instead, we create a new struct,
`merge::inputs::CherryPick`, that implements the same new trait
`merge::inputs::MergeInputs` as `merge::inputs::Inputs`, but its fields
are populated "by hand" so we can select E^ as the base commit. This
object can then be passed to `merge::resolve::Resolve` to perform the
merge.

That gets the index and workspace into the right state, and we then
write a commit. This commit reuses the author and message from the
picked commit, but sets the current author for the `committer` field.
The commit also has a single parent, unlike true merges, which have two.