[23.2.1] Cherry-pick multiple commits and ranges
Our current `cherry-pick` implementation accepts a single commit. We can
extend it by allowing it to take multiple arguments, and applying all
the requested commits in sequence. This allows for the picking of
ranges, and will also let the whole sequence be rolled back if a problem
is encountered part-way through.

When multiple arguments are given that are single commits, as in
`cherry-pick A B C`, then each commit should be applied in the order
given. If any of the arguments are ranges, as in `cherry-pick A..B C`,
then the arguments are interpreted by `rev-list`, so this picks all
commits reachable from A or C but not from B, in the order given by
`rev-list`.

To support this behaviour `RevList` needs a mode where it can take this
list of arguments and determine if any of them represent ranges or
excluded commits as in ^B. If none of them are ranges then its behaviour
should change in two ways:

- The order of the arguments should be preserved; they should not be
  sorted by `committer` date as they usually are.

- We should yield only the given commits, not any of their ancestors.

In Git this mode is triggered by the `--no-walk` option, which we'll
replicate by passing `walk: false`. If this option is used, the above
behaviour is respected. If any of the inputs represent excluded commits,
then the `walk` option is forced back on.

`RevList` yields commits in reverse-chronological order, whereas it
makes more sense to cherry-pick commits in forwards order. So,
`CherryPick` iterates over the commits in reverse, and reverses the
arguments passed into `RevList` so their order is effectively preserved
in the case where no ranges are used.