[24.3.2] Extract sequencing code in preparation for `revert`
We'll shortly add the `revert` command, which works very much like the
`cherry-pick` command in that it takes a sequence of commits and
performs a kind of merge with each one. It has the same machinery as
`cherry-pick` for dealing with failures: `--continue`, `--abort` and
`--quit` options, and state stored in `.git/sequencer` and
`.git/REVERT_HEAD` for recovery.

`revert` will differ in how it selects the initial commits, and how it
constructs each commit after merging, but all the state management
functionality is the same. Much of this shared behaviour is currently in
the `CherryPick` class, and by extracting it into
`commands::shared::sequencing` we can reuse it with `revert`.

Note that this doesn't quite follow the book, which introduces a
`Sequencing` class with all of this behavior. It also deviates from the
book by leaving `handle_continue()` on the `CherryPick` command because
extracting that makes the borrow checker very sad in ways I was unable
to resolve.