[24.3.3] Fully working `revert` command, based on `cherry-pick`
After putting the infrastructure in place in recent commits, we can now
introduce the `revert` command. This is essentially the same as
`cherry-pick` in structure, but performs a different action for each
commit: instead of applying that commit's changes, it applies the
_reverse_ change, effectively undoing the commit. For example, given the
following history:

    A   B   C   D
    o---o---o---o

Reverting C will calculate the change between C and B (i.e. the inverse
of the change in commit C), and apply this diff on top of D. If D has
changed some of the same files as C, a conflict may result.

`revert` can be implemented by reusing much of the `cherry-pick` code
and only implementing new functionality where it differs from that
command. Those differences are as follows.

The ID of any pending commits is stored in `.git/REVERT_HEAD` to
distinguish it from cherry-picks and true merges. This is triggered by
using the commit-type `Revert` with `PendingCommit`.

The initial commit sequence is stored in reverse chronological order
when ranges are used, i.e. the default order of `RevList`. This is
because when undoing changes, it makes more sense to remove them in the
reverse order to that in which they were applied. Not only does this
make logical sense; changes will not necessarily commute if you try to
apply their inverses in the same order, whereas they will commute if you
apply them in reverse order.

The inputs passed to `Resolve` express the idea of taking the diff from
the reverted commit to its parent: the reverted commit becomes the base
of the merge, and its parent becomes the right-hand input.

The final commit does not reuse the author or message from the original
commit, because we're not reproducing it, we're undoing it. Instead the
default message states which commit is being reverted. Unlike
`cherry-pick`, `revert` lets this message be edited by default.

When `commit` is run with `.git/REVERT_HEAD` existing, a revert commit
is written from the message stashed in `.git/MERGE_MSG`.