[24.3.4] Show pending commit state in `status`
It's useful to be able to see the state of any pending commit when the
`status` command is run. After showing the name of the branch, we can
show a message showing which kind of merge is in progress: a
cherry-pick, a revert, or a true merge, with different instructions
depending on whether any conflicts are still present.