[25.2] Introduce `Config` class for managing configuration files
Git stores configuration in various places: `.git/config`,
`~/.gitconfig`, `/etc/gitconfig`, and also uses the configuration format
for storing other pieces of state, for example `.git/sequencer/opts`.
This `Config` class implements the file format for each of these files,
and presents an interface for programmatically updating and reading
their contents.

Config keys are presented in parsed form to this class, for example the
key that appears to the user as `core.editor` becomes `["core",
"editor"]`. This means internal uses of `Config` aren't concerned with
parsing keys, making it safer to use key segments that themselves
contain dots.

The class supports the following methods:

- `open`, `open_for_update`: these either simply read the file into
  memory, or do so after writing a lockfile, giving them a write lock
  before making updates.

- `save`: this writes the updated config to disk, removing the lock.

- `get`: returns the last value of a key. Some values like numbers and
  booleans are automatically converted from strings.

- `get_all`: returns all the values of a key, in source order.

- `add`: adds a new value for a key, allowing multiple values for the
  same key to exist in a file.

- `set`: sets a single value for a key; the key must have at most one
  existing value for this to work.

- `replace_all`: should be used in place of `set` if your intention is
  to remove a multi-valued key and replace it with a single value.

The parser is written in a way that attempts to preserve formatting, so
any comments, blank lines and indentation will be preserved on existing
lines. Newly added variables will be appended at the end of any existing
section as appropriate, and use a tab for indentation.