[25.3.2] Use the `user.{name,email}` to set author information
Until we introduced support for configuration, we used the
`GIT_AUTHOR_NAME` and `GIT_AUTHOR_EMAIL` environment variables to set
the author information for new commits. Now that we can read from
configuration files, we can use the `user.name` and `user.email`
variables as the sources for this information.